requirements.txt
PyMuPDF>=1.23.0
python-docx>=1.1.0
pypdf>=4.0.0
Pillow>=10.0.0
tqdm>=4.66.0
flask>=3.0.0
scikit-image>=0.24.0
google-generativeai>=0.8.0
